Move script/style logic into script-loader.php
